feat: update README, LICENSE, add Contrib, CoC
This PR improves the overall FOSS friendliness of the project.
It improves the README structure,
corrects copyright in LICENSE, adds a general Contribution
and a standard code of conduct.
The code of conduct  uses the well used
Contributor Covenant, https://www.contributor-covenant.org/.
The CONTRIBUTING is based on a template taken from
Digg's Open Source Guild, and should been seen as
a help base, that with time, can be adjusted to fit the project better,
if needed.

Signed-off-by: Josef Andersson <josef.andersson@digg.se>