Make project REUSE Compliant
This PR makes the project follow the REUSE standard.
Basically, it definies the standard SPDX license header,
and adds "general or binary" license declarations in the .resuse/dep5 file (as per REUSE spec).

In short: it tells the licensing of every asset in your project.

It also adds a workflow that checks license compliance before the test step.
The first step would be ideal to also add lints etc in the future.

Fixes #14

Signed-off-by: Josef Andersson <josef.andersson@digg.se>